+++ bfabric/src/bfabric/bfabric2.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-import warnings
-
-warnings.warn("bfabric.bfabric2 module is deprecated, use bfabric instead", DeprecationWarning)
-# TODO deprecated - import from bfabric instead