+++ docs/changelog.md
@@ -10,9 +10,16 @@ Versioning currently follows `X.Y.Z` where
 
 ## \[Unreleased\]
 
+## \[1.13.20\] - 2025-02-10
+
+### Breaking
+
+- the old `bfabric.cli_formatting` and `bfabric.bfabric2` modules have been deleted
+
 ### Changed
 
 - move `use_client` to `bfabric.utils.cli_integration`: this will allow reuse between bfabric-scripts and bfabric-app-runner
+- move functionality from `bfabric.cli_formatting` to `bfabric.utils.cli_integration`
 
 ## \[1.13.19\] - 2025-02-06